+++ src/main/java/domain/card/CardPattern.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
