+++ src/main/java/dto/MemberStatus.java
@@ -1,6 +1,6 @@
-package domain.dto;
+package dto;
 
-import domain.Card;
+import domain.card.Card;
 import java.util.List;
 
 public record MemberStatus(
