+++ src/test/java/domain/card/CardNumberTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
