+++ src/test/java/domain/card/FixedDeck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.LinkedList;
 import java.util.List;
